fix non blocking
